+++ utils.c
@@ -6,13 +6,13 @@
 /*   By: jbakker <jbakker@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/12/07 22:07:15 by jbakker       #+#    #+#                 */
-/*   Updated: 2024/12/10 02:38:17 by jbakker       ########   odam.nl         */
+/*   Updated: 2024/12/11 16:07:58 by jbakker       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cube3d.h"
 
-double	min(double a, double b)
+float	min(float a, float b)
 {
 	if (a < b)
 		return (a);
@@ -31,12 +31,12 @@ int	rgb_to_int(int r, int g, int b)
 	return (r << 16 | g << 8 | b);
 }
 
-double	distance(double x1, double y1, double x2, double y2)
+float	distance(float x1, float y1, float x2, float y2)
 {
 	return (sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1)));
 }
 
-double	normalize_angle(double angle)
+float	normalize_angle(float angle)
 {
 	while (angle > 2 * PI)
 		angle -= 2 * PI;
